# api/openapi/v1/vertera.yaml
openapi: 3.1.0
info:
  title: Vertera Public API
  version: 1.0.0
  description: |
    Vertera REST API v1. All endpoints are prefixed with /api/v1.
    Authentication: OIDC Bearer tokens (JWT). Agents use mTLS gRPC, not this API.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /projects:
    get:
      tags: [Projects]
      summary: List projects
      operationId: listProjects
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/pageSize'
      responses:
        '200':
          description: OK
  /agents/enroll/token:
    post:
      tags: [Agents]
      summary: Issue a short-lived enrollment token (HS256-signed)
      operationId: createEnrollToken
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EnrollTokenRequest' }
      responses:
        '200':
          description: Token issued
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EnrollTokenResponse' }
  /agents/enroll/csr:
    post:
      tags: [Agents]
      summary: Submit a CSR and receive a signed client certificate
      operationId: signAgentCSR
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CsrSignRequest' }
      responses:
        '200':
          description: Signed certificate
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CsrSignResponse' }
...
  /hosts/{hostId}/packages/install:
    post:
      tags: [Hosts]
      summary: Install packages (OVS/CH)
      operationId: installPackages
      parameters:
        - $ref: '#/components/parameters/hostId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PackageInstallRequest' }
      responses:
        '202':
          description: Task accepted
          headers:
            Location:
              description: URL to poll task status
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Task' }
...
  /tasks/{taskId}:
    parameters:
      - $ref: '#/components/parameters/taskId'
    get:
      tags: [Tasks]
      summary: Get task
      operationId: getTask
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Task' }
  /tokens/enroll:
    post:
      tags: [Tokens]
      summary: Create a short-lived seed enrollment token
      operationId: createEnrollToken
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EnrollTokenCreate' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EnrollToken' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    page:
      in: query
      name: page
      schema: { type: integer, minimum: 1, default: 1 }
    pageSize:
      in: query
      name: pageSize
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
...
    hostId:
      in: path
      name: hostId
      required: true
      schema: { type: string, format: uuid }
...
    taskId:
      in: path
      name: taskId
      required: true
      schema: { type: string, format: uuid }
  schemas:
...
    PackageInstallRequest:
      type: object
      properties:
        packages:
          type: array
          items: { type: string, enum: [ovs, cloud-hypervisor] }
        version: { type: string }
...
    Task:
      type: object
      properties:
        id: { type: string }
        hostId: { type: string }
        type: { type: string }
        params: { type: object, additionalProperties: true }
        status: { type: string, enum: [queued, running, succeeded, failed] }
        logs: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
        startedAt: { type: string, format: date-time, nullable: true }
        finishedAt: { type: string, format: date-time, nullable: true }
        error: { type: string, nullable: true }
    EnrollTokenCreate:
      type: object
      required: [projectId]
      properties:
        projectId: { type: string, format: uuid }
        siteId: { type: string, format: uuid, nullable: true }
        os_version: { type: string, description: EL version, example: el9 }
    EnrollTokenRequest:
      type: object
      properties:
        ttl:
          type: string
          description: Token TTL in Go duration (e.g., 15m, 1h). Defaults to 15m.
    EnrollTokenResponse:
      type: object
      properties:
        token: { type: string }
        expires_at: { type: string, format: date-time }
    CsrSignRequest:
      type: object
      required: [token, csr_pem]
      properties:
        token: { type: string }
        csr_pem: { type: string, description: PEM-encoded PKCS#10 CSR }
    CsrSignResponse:
      type: object
      properties:
        cert_pem: { type: string, description: PEM-encoded X.509 certificate }
    EnrollToken:
      type: object
      properties:
        id: { type: string, format: uuid }
        token: { type: string, description: Signed JWT }
        expiresAt: { type: string, format: date-time }
        oneTime: { type: boolean, default: true }
        useLimit: { type: integer, default: 1 }
        constraints:
          type: object
          properties:
            allowedCidrs: { type: array, items: { type: string } }
            requireTpm: { type: boolean }
            modelAllowlist: { type: array, items: { type: string } }